requirements.txt
aiogram==3.20.0.post0
aiohttp==3.11.18
asyncpg==0.30.0
python-dotenv==1.1.1
fastapi==0.115.12
uvicorn[standard]==0.34.0
jinja2==3.1.6
python-multipart==0.0.20
itsdangerous==2.2.0
